[fix] : 인기 장학금 조회 시 좋아요수 필드 타입 수정

+++ src/main/java/backend/univfit/domain/apply/api/dto/response/PopularAnnouncementResponse.java
@@ -6,10 +6,10 @@ public record PopularAnnouncementResponse(
         String scholarShipName,
         String scholarshipFoundation,
         String applicationPeriod,
-        Integer howManyLikes
+        Long howManyLikes
 ) {
     public static PopularAnnouncementResponse of(Long announcementId, String scholarShipImage, String scholarShipName, String scholarshipFoundation,
-                                                 String applicationPeriod, Integer howManyLikes) {
+                                                 String applicationPeriod, Long howManyLikes) {
         return new PopularAnnouncementResponse(announcementId, scholarShipImage, scholarShipName, scholarshipFoundation,
                 applicationPeriod, howManyLikes);
     }
